Import mozilla 1.7.
git-svn-id: svn+ssh://svn.mit.edu/athena/branches/vendor@20550 728af825-273c-0410-89f9-f7d3b574a069

+++ third/mozilla/extensions/wallet/cookieviewer/resources/content/CookieViewer.js
@@ -27,6 +27,7 @@ var kObserverService;
 // interface variables
 var cookiemanager         = null;          // cookiemanager interface
 var permissionmanager     = null;          // permissionmanager interface
+var promptservice         = null;          // promptservice interface
 var popupmanager          = null;          // popup manager
 var gDateService = null;
 
@@ -38,9 +39,11 @@ var deletedPermissions   = [];
 
 // differentiate between cookies, images, and popups
 const nsIPermissionManager = Components.interfaces.nsIPermissionManager;
+const nsICookiePermission = Components.interfaces.nsICookiePermission;
 const cookieType = "cookie";
 const imageType = "image";
 const popupType = "popup";
+
 
 var dialogType = cookieType;
 if (window.arguments[0] == "imageManager")
@@ -49,6 +52,7 @@ else if (window.arguments[0] == "popupManager")
   dialogType = popupType;
 
 var cookieBundle;
+var gUpdatingBatch = "";
 
 function Startup() {
 
@@ -57,11 +61,13 @@ function Startup() {
   //   cookieManagerFromIcon
   //   imageManager
  
-  // xpconnect to cookiemanager/permissionmanager/popupmanager interfaces
+  // xpconnect to cookiemanager/permissionmanager/promptservice/popupmanager interfaces
   cookiemanager = Components.classes["@mozilla.org/cookiemanager;1"].getService();
   cookiemanager = cookiemanager.QueryInterface(Components.interfaces.nsICookieManager);
   permissionmanager = Components.classes["@mozilla.org/permissionmanager;1"].getService();
   permissionmanager = permissionmanager.QueryInterface(Components.interfaces.nsIPermissionManager);
+  promptservice = Components.classes["@mozilla.org/embedcomp/prompt-service;1"].getService();
+  promptservice = promptservice.QueryInterface(Components.interfaces.nsIPromptService);
   popupmanager = Components.classes["@mozilla.org/PopupWindowManager;1"].getService();
   popupmanager = popupmanager.QueryInterface(Components.interfaces.nsIPopupWindowManager);
 
@@ -96,6 +102,8 @@ function Startup() {
       element.value = cookieBundle.getString("textBannedImages");
       element = document.getElementById("cookiesTab");
       element.hidden = "true";
+      element = document.getElementById("btnSession");
+      element.hidden = "true";
     } else {
       element = document.getElementById("cookieviewer");
       element.setAttribute("title", cookieBundle.getString("popupTitle"));
@@ -109,7 +117,6 @@ function Startup() {
     }
   } catch(e) {
   }
-
   // load in the cookies and permissions
   cookiesTree = document.getElementById("cookiesTree");
   permissionsTree = document.getElementById("permissionsTree");
@@ -131,6 +138,8 @@ function Shutdown() {
 
 var cookieReloadDisplay = {
   observe: function(subject, topic, state) {
+    if (topic == gUpdatingBatch)
+      return;
     if (topic == "cookieChanged") {
       if (state == "cookies") {
         cookies.length = 0;
@@ -161,12 +170,19 @@ var cookiesTreeView = {
   getCellValue : function(row,column) {},
   getCellText : function(row,column){
     var rv="";
-    if (column=="domainCol") {
+    switch (column) {
+    case "domainCol":
       rv = cookies[row].rawHost;
-    } else if (column=="nameCol") {
+      break;
+    case "nameCol":
       rv = cookies[row].name;
-    } else if (column=="statusCol") {
-      rv = GetStatusString(cookies[row].status);
+      break;
+    case "statusCol":
+      rv = cookies[row].status;
+      break;
+    case "expiresCol":
+      rv = cookies[row].expires;
+      break;
     }
     return rv;
   },
@@ -190,8 +206,9 @@ function Cookie(number,name,value,isDomain,host,rawHost,path,isSecure,expires,
   this.rawHost = rawHost;
   this.path = path;
   this.isSecure = isSecure;
-  this.expires = expires;
-  this.status = status;
+  this.expires = GetExpiresString(expires);
+  this.expiresSortValue = expires;
+  this.status = GetStatusString(status);
   this.policy = policy;
 }
 
@@ -257,10 +274,21 @@ function loadCookies() {
 function GetExpiresString(expires) {
   if (expires) {
     var date = new Date(1000*expires);
-    return gDateService.FormatDateTime("", gDateService.dateFormatLong,
-                                       gDateService.timeFormatSeconds, date.getFullYear(),
-                                       date.getMonth()+1, date.getDate(), date.getHours(),
-                                       date.getMinutes(), date.getSeconds());
+
+    // if a server manages to set a really long-lived cookie, the dateservice
+    // can't cope with it properly, so we'll just return a blank string
+    // see bug 238045 for details
+    var expiry = "";
+    try {
+      expiry = gDateService.FormatDateTime("", gDateService.dateFormatLong,
+                                           gDateService.timeFormatSeconds, 
+                                           date.getFullYear(), date.getMonth()+1, 
+                                           date.getDate(), date.getHours(),
+                                           date.getMinutes(), date.getSeconds());
+    } catch(ex) {
+      // do nothing
+    }
+    return expiry;
   }
   return cookieBundle.getString("AtEndOfSession");
 }
@@ -298,6 +326,7 @@ function CookieSelected() {
   if (selections.length) {
     document.getElementById("removeCookie").removeAttribute("disabled");
   } else {
+    document.getElementById("removeCookie").setAttribute("disabled", "true");
     ClearCookieProperties();
     return true;
   }
@@ -307,7 +336,7 @@ function CookieSelected() {
     // Something got out of synch.  See bug 119812 for details
     dump("Tree and viewer state are out of sync! " +
          "Help us figure out the problem in bug 119812");
-    return;
+    return false;
   }
 
   var props = [
@@ -320,8 +349,9 @@ function CookieSelected() {
     {id: "ifl_path", value: cookies[idx].path},
     {id: "ifl_isSecure",
      value: cookies[idx].isSecure ?
-            cookieBundle.getString("yes") : cookieBundle.getString("no")},
-    {id: "ifl_expires", value: GetExpiresString(cookies[idx].expires)},
+            cookieBundle.getString("forSecureOnly") : 
+            cookieBundle.getString("forAnyConnection")},
+    {id: "ifl_expires", value: cookies[idx].expires},
     {id: "ifl_policy", value: GetPolicyString(cookies[idx].policy)}
   ];
 
@@ -359,6 +389,11 @@ function DeleteCookie() {
 }
 
 function DeleteAllCookies() {
+  var title = cookieBundle.getString("deleteAllCookiesTitle");
+  var msg = cookieBundle.getString("deleteAllCookies");
+  if (!promptservice.confirm(window, title, msg ))
+    return;
+
   ClearCookieProperties();
   DeleteAllFromTree(cookiesTree, cookiesTreeView,
                         cookies, deletedCookies,
@@ -367,6 +402,7 @@ function DeleteAllCookies() {
 }
 
 function FinalizeCookieDeletions() {
+  gUpdatingBatch = "cookieChanged";
   for (var c=0; c<deletedCookies.length; c++) {
     cookiemanager.remove(deletedCookies[c].host,
                          deletedCookies[c].name,
@@ -374,6 +410,7 @@ function FinalizeCookieDeletions() {
                          document.getElementById("checkbox").checked);
   }
   deletedCookies.length = 0;
+  gUpdatingBatch = "";
 }
 
 function HandleCookieKeyPress(e) {
@@ -387,9 +424,38 @@ var lastCookieSortAscending = false;
 
 function CookieColumnSort(column) {
   lastCookieSortAscending =
-    SortTree(cookiesTree, cookiesTreeView, cookies,
-                 column, lastCookieSortColumn, lastCookieSortAscending);
+      SortTree(cookiesTree, cookiesTreeView, cookies,
+               column, lastCookieSortColumn, lastCookieSortAscending);
   lastCookieSortColumn = column;
+  // set the sortDirection attribute to get the styling going
+  // first we need to get the right element
+  var sortedCol;
+  switch (column) {
+    case "rawHost":
+      sortedCol = document.getElementById("domainCol");
+      break;
+    case "name":
+      sortedCol = document.getElementById("nameCol");
+      break;
+    case "expires":
+      sortedCol = document.getElementById("expiresCol");
+      break;
+    case "status":
+      sortedCol = document.getElementById("statusCol");
+      break;
+  }
+  if (lastCookieSortAscending)
+    sortedCol.setAttribute("sortDirection", "descending");
+  else
+    sortedCol.setAttribute("sortDirection", "ascending");
+
+  // clear out the sortDirection attribute on the rest of the columns
+  var currentCol = sortedCol.parentNode.firstChild;
+  while (currentCol) {
+    if (currentCol != sortedCol && currentCol.localName == "treecol")
+      currentCol.removeAttribute("sortDirection");
+    currentCol = currentCol.nextSibling;
+  }
 }
 
 /*** =================== PERMISSIONS CODE =================== ***/
@@ -404,7 +470,7 @@ var permissionsTreeView = {
     var rv="";
     if (column=="siteCol") {
       rv = permissions[row].rawHost;
-    } else if (column=="statusCol") {
+    } else if (column=="capabilityCol") {
       rv = permissions[row].capability;
     }
     return rv;
@@ -432,28 +498,49 @@ function loadPermissions() {
   var enumerator = permissionmanager.enumerator;
   var count = 0;
   var contentStr;
-  var canStr, cannotStr;
-  if (dialogType == cookieType) {
-    canStr="can";
-    cannotStr="cannot";
-  } else if (dialogType == imageType) {
-    canStr="canImages";
-    cannotStr="cannotImages";
-  } else {
-    canStr="canPopups";
-    cannotStr="cannotPopups";
+  var canStr, cannotStr, canSessionStr;
+  switch (dialogType) {
+    case cookieType:
+      canStr="can";
+      canSessionStr="canSession";
+      cannotStr="cannot";
+      break;
+    case imageType:
+      canStr="canImages";
+      cannotStr="cannotImages";
+      break;
+    case popupType:
+      canStr="canPopups";
+      cannotStr="cannotPopups";
+      break;
+    default:
+      return;
   }
   while (enumerator.hasMoreElements()) {
     var nextPermission = enumerator.getNext();
     nextPermission = nextPermission.QueryInterface(Components.interfaces.nsIPermission);
     if (nextPermission.type == dialogType) {
       var host = nextPermission.host;
-      var capability = ( nextPermission.capability == nsIPermissionManager.ALLOW_ACTION ) ? true : false;
+      var capability;
+      switch (nextPermission.capability) {
+        case nsIPermissionManager.ALLOW_ACTION:
+          capability = canStr;
+          break;
+        case nsIPermissionManager.DENY_ACTION:
+          capability = cannotStr;
+          break;
+        case nsICookiePermission.ACCESS_SESSION:
+          // we should only hit this for cookies
+          capability = canSessionStr;
+          break;
+        default:
+          return;
+      }
       permissions[count] = 
         new Permission(count++, host,
                        (host.charAt(0)==".") ? host.substring(1,host.length) : host,
                        nextPermission.type,
-                       cookieBundle.getString(capability?canStr:cannotStr));
+                       cookieBundle.getString(capability));
     }
   }
   permissionsTreeView.rowCount = permissions.length;
@@ -464,7 +551,7 @@ function loadPermissions() {
 
   // disable "remove all" button if there are no cookies/images
   if (permissions.length == 0) {
-    document.getElementById("removeAllPermissions").setAttribute("disabled","true");
+    document.getElementById("removeAllPermissions").setAttribute("disabled", "true");
   } else {
     document.getElementById("removeAllPermissions").removeAttribute("disabled");
   }
@@ -473,9 +560,10 @@ function loadPermissions() {
 
 function PermissionSelected() {
   var selections = GetTreeSelections(permissionsTree);
-  if (selections.length) {
+  if (selections.length)
     document.getElementById("removePermission").removeAttribute("disabled");
-  }
+  else
+    document.getElementById("removePermission").setAttribute("disabled", "true");
 }
 
 function DeletePermission() {
@@ -492,7 +580,10 @@ function setCookiePermissions(action) {
                             .getService(Components.interfaces.nsIIOService);
   var uri = ioService.newURI(url, null, null);
   
-  permissionmanager.add(uri, dialogType, action);
+  // only set the permission if the permission doesn't already exist
+  if (permissionmanager.testPermission(uri, dialogType) != action)
+    permissionmanager.add(uri, dialogType, action);
+
   site.focus();
   site.value = "";
 }
@@ -501,12 +592,21 @@ function buttonEnabling(textfield) {
   // trim any leading space
   var site = textfield.value.replace(/^\s*([-\w]*:\/+)?/, "");
   var block = document.getElementById("btnBlock");
+  var session = document.getElementById("btnSession");
   var allow = document.getElementById("btnAllow");
   block.disabled = !site;
+  if (dialogType == cookieType)
+    session.disabled = !site;
   allow.disabled = !site;
 }
 
 function DeleteAllPermissions() {
+  var title = cookieBundle.getString("deleteAllSitesTitle");
+  var msgType = dialogType == cookieType ? "deleteAllCookiesSites" : "deleteAllImagesSites";
+  var msg = cookieBundle.getString(msgType);
+  if (!promptservice.confirm(window, title, msg ))
+    return;
+    
   DeleteAllFromTree(permissionsTree, permissionsTreeView,
                         permissions, deletedPermissions,
                         "removePermission", "removeAllPermissions");
@@ -514,19 +614,24 @@ function DeleteAllPermissions() {
 }
 
 function FinalizePermissionDeletions() {
-  var ioService = Components.classes["@mozilla.org/network/io-service;1"]
-                    .getService(Components.interfaces.nsIIOService);
+  if (!deletedPermissions.length)
+    return;
 
-  for (var p=0; p<deletedPermissions.length; p++) {
-    if (deletedPermissions[p].type == popupType) {
+  gUpdatingBatch = "perm-changed";
+  var p;
+  if (deletedPermissions[0].type == popupType) {
+    var ioService = Components.classes["@mozilla.org/network/io-service;1"]
+                              .getService(Components.interfaces.nsIIOService);
+    for (p = 0; p < deletedPermissions.length; ++p)
       // we lost the URI's original scheme, but this will do because the scheme
       // is stripped later anyway.
-      var uri = ioService.newURI("http://"+deletedPermissions[p].host, null, null);
-      popupmanager.remove(uri);
-    } else
+      popupmanager.remove(ioService.newURI("http://"+deletedPermissions[p].host, null, null));
+  } else {
+    for (p = 0; p < deletedPermissions.length; ++p)
       permissionmanager.remove(deletedPermissions[p].host, deletedPermissions[p].type);
   }
   deletedPermissions.length = 0;
+  gUpdatingBatch = "";
 }
 
 function HandlePermissionKeyPress(e) {
@@ -544,6 +649,29 @@ function PermissionColumnSort(column, updateSelection) {
                  column, lastPermissionSortColumn, lastPermissionSortAscending, 
                  updateSelection);
   lastPermissionSortColumn = column;
+  
+  // make sure sortDirection is set
+  var sortedCol;
+  switch (column) {
+    case "rawHost":
+      sortedCol = document.getElementById("siteCol");
+      break;
+    case "capability":
+      sortedCol = document.getElementById("capabilityCol");
+      break;
+  }
+  if (lastPermissionSortAscending)
+    sortedCol.setAttribute("sortDirection", "descending");
+  else
+    sortedCol.setAttribute("sortDirection", "ascending");
+
+  // clear out the sortDirection attribute on the rest of the columns
+  var currentCol = sortedCol.parentNode.firstChild;
+  while (currentCol) {
+    if (currentCol != sortedCol && currentCol.localName == "treecol")
+      currentCol.removeAttribute("sortDirection");
+    currentCol = currentCol.nextSibling;
+  }
 }
 
 /*** ============ CODE FOR HELP BUTTON =================== ***/
